fix(build-admission): stop stale lifecycle rows denying capacity they never held
Startup recovery still compared `count_task_or_warm_occupancy()` -- a count of
occupying JOURNAL rows -- against `self.cap`, a count of BUILD SLOTS. Those were
already two different units, and once the journal became the lifecycle ledger
rather than a capacity authority the comparison stopped meaning anything at all:
a journal row records that an object existed, it does not reserve CPU.

This is not academic. It is the production wedge seen on v0.7.5: 59 occupying
journal rows, 58 of them stale from a predecessor whose objects are long gone,
against a cap of 3. That latched `over_cap`, which `readiness()` reports as
SeededOccupancyAboveCap, which fails Enforce closed for every admission. Stale
lifecycle rows are reconciliation's job (#2597); they must never have been able
to deny capacity they were not holding.

The gate now reads the one capacity authority. No authority installed means the
controller is not capacity-gated, so there is nothing to exceed; an authority
that IS installed but unreadable leaves the gate closed, because unknown
occupancy is not proof of headroom. The now-dead journal count is removed rather
than left assigned, so nothing invites comparing it to a cap again.

Also re-labels the handoff contract that this cut-over changed: V0Mode::Enforce
means the LEDGER is authoritative and fails closed, not that v0 enforces a cap
(it has none). validate_admission_config keeps its "neither authority enforces"
rule under that reading; only V1Mode::Enforce arms the build-slot cap.

Splits the warmer ledger test into its own file: graph_warmer_tests.rs sat 655
bytes under the 51200-byte Server Guards ceiling, which is a trap for the next
author. No allow-oversize marker added; the guard reports 0 unmarked files.
