+++ server/crates/djinn-k8s/src/graph_warmer_tests.rs
@@ -199,7 +199,7 @@ fn watch_deadline_follows_configured_job_timeout_plus_slack() {
 /// columns are no longer consulted. So we create a catalog `images` row,
 /// mark it ready with the expected content-addressed tag (no digest, so
 /// `pull_ref` returns the tag verbatim), and assign it to the project.
-async fn seed_project_with_ready_image(db: &Database, name: &str) -> String {
+pub(super) async fn seed_project_with_ready_image(db: &Database, name: &str) -> String {
     use djinn_db::ImageRepository;
     let repo = ProjectRepository::new(db.clone(), EventBus::noop());
     let project = repo
@@ -228,7 +228,7 @@ async fn seed_project_with_ready_image(db: &Database, name: &str) -> String {
     project.id
 }
 
-fn test_config() -> KubernetesConfig {
+pub(super) fn test_config() -> KubernetesConfig {
     KubernetesConfig::for_testing()
 }
 
@@ -1058,10 +1058,10 @@ async fn trigger_skips_completion_sink_when_job_disappears() {
     );
 }
 
-struct AdmissionRecording {
-    events: Arc<Mutex<Vec<String>>>,
-    admit_error: Option<WarmAdmissionError>,
-    fail_create_started: bool,
+pub(super) struct AdmissionRecording {
+    pub(super) events: Arc<Mutex<Vec<String>>>,
+    pub(super) admit_error: Option<WarmAdmissionError>,
+    pub(super) fail_create_started: bool,
 }
 
 #[async_trait]
@@ -1096,9 +1096,9 @@ impl WarmAdmission for AdmissionRecording {
     }
 }
 
-struct LifecycleRecordingDispatcher {
-    events: Arc<Mutex<Vec<String>>>,
-    result: Result<String, String>,
+pub(super) struct LifecycleRecordingDispatcher {
+    pub(super) events: Arc<Mutex<Vec<String>>>,
+    pub(super) result: Result<String, String>,
 }
 
 #[async_trait]
@@ -1109,7 +1109,7 @@ impl WarmJobDispatcher for LifecycleRecordingDispatcher {
     }
 }
 
-struct UidWatcher;
+pub(super) struct UidWatcher;
 
 #[async_trait]
 impl WarmJobWatcher for UidWatcher {
@@ -1316,83 +1316,3 @@ fn dispatcher_error_classification_is_conservative() {
 #[path = "graph_warmer_overlap_tests.rs"]
 mod overlap;
 
-/// A lease that always grants, standing in for an armed v1 build lease.
-struct AlwaysGrantingLease;
-
-#[async_trait]
-impl crate::GraphWarmLease for AlwaysGrantingLease {
-    async fn acquire(
-        &self,
-        identity: djinn_supervisor::services::GraphWarmLeaseIdentity,
-        _deadlines: djinn_supervisor::services::LeaseDeadlines,
-    ) -> Result<crate::GraphWarmLeaseGrant, crate::GraphWarmLeaseError> {
-        Ok(crate::GraphWarmLeaseGrant {
-            identity,
-            grant: djinn_supervisor::services::LeaseGrant {
-                fencing_token: LeaseFencingToken(7),
-                deadlines: djinn_supervisor::services::LeaseDeadlines {
-                    queue_deadline_ms: 0,
-                    launch_deadline_ms: 0,
-                },
-            },
-        })
-    }
-
-    async fn bind(
-        &self,
-        _identity: &djinn_supervisor::services::GraphWarmLeaseIdentity,
-        _fencing_token: LeaseFencingToken,
-        _pod_uid: String,
-    ) -> Result<(), crate::GraphWarmLeaseError> {
-        Ok(())
-    }
-}
-
-/// A lease-granted warm Job must STILL be recorded in the lifecycle ledger.
-///
-/// This drives the warmer itself, not just the controller, because the defect
-/// lived in the warmer's own match: `(Some(lease_grant), _) => None` meant that
-/// once the v1 lease was armed -- which is always, in production, whenever the
-/// cap is positive -- a warm Job wrote no admission-journal row at all. The
-/// journal is what records the Kubernetes UID and what #2597's
-/// `reclaim_absent_object` reads to retire occupancy whose object is provably
-/// gone, so a leased warm Job was invisible to reclamation.
-///
-/// The lease still governs whether the Job may be created; the ledger append is
-/// capacity-neutral and simply must also happen.
-#[tokio::test]
-async fn a_lease_granted_warm_job_is_still_recorded_in_the_lifecycle_ledger() {
-    let db = Database::open_in_memory().expect("in-memory db");
-    let project_id = seed_project_with_ready_image(&db, "proj-leased-ledger").await;
-    let events = Arc::new(Mutex::new(Vec::new()));
-    let warmer = K8sGraphWarmer::with_dispatcher(
-        test_config(),
-        db,
-        Arc::new(LifecycleRecordingDispatcher {
-            events: events.clone(),
-            result: Ok("leased-warm".to_string()),
-        }),
-        Arc::new(UidWatcher),
-    )
-    .with_graph_warm_lease(Arc::new(AlwaysGrantingLease))
-    .with_warm_admission(Arc::new(AdmissionRecording {
-        events: events.clone(),
-        admit_error: None,
-        fail_create_started: false,
-    }));
-
-    warmer.trigger(&project_id).await;
-    tokio::task::yield_now().await;
-    tokio::time::sleep(Duration::from_millis(50)).await;
-
-    let events = events.lock().await;
-    assert!(
-        events.iter().any(|event| event.starts_with("reserve:")),
-        "a lease-granted warm Job wrote no ledger row: {events:?} \
-         -- it is invisible to absent-object reclamation"
-    );
-    assert!(
-        events.iter().any(|event| event == "CreateStarted"),
-        "the ledger row must reach CreateStarted before the POST: {events:?}"
-    );
-}
